+++ config/initializers/fuzzy_where_config.rb
@@ -1,4 +1,5 @@
 FuzzyWhere.configure do |config|
   # config.where_method_name = :fuzzy_where
+  # config.membership_degree_column_name = :membership_degree
   config.predicates_file = Rails.root.join('config', 'fuzzy_predicates.yml')
 end