+++ spec/fuzzy_where/fuzzy_derivation_spec.rb
@@ -2,8 +2,8 @@ require 'spec_helper'
 
 if defined? ActiveRecord
   describe FuzzyWhere::FuzzyDerivation do
-    describe "#derivative_query" do
-      skip "Pending test"
+    describe '#derivative_conditions' do
+      skip 'Pending test'
     end
   end
 end