+++ spec/fuzzy_where/fuzzy_relation_builder_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+describe FuzzyWhere::FuzzyRelationBuilder do
+  describe '#build' do
+    skip 'Pending test'
+  end
+end