+++ spec/support/active_record/ar_stand_in.rb
@@ -1,4 +1,4 @@
-class ARStandIn < ActiveRecord::Base;
+class ARStandIn < ActiveRecord::Base
   self.abstract_class = true
   include FuzzyWhere::ActiveRecordModelExtension
 end