+++ spec/support/active_record/hotel.rb
@@ -1,3 +1,2 @@
 class Hotel < ActiveRecord::Base
-
 end