+++ spec/support/active_record/migrations.rb
@@ -1,4 +1,4 @@
-#migrations
+# Migrations
 class CreateAllTables < ActiveRecord::Migration
   def self.up
     create_table(:people) { |t| t.string :name; t.integer :age }