+++ spec/support/active_record/person.rb
@@ -1,3 +1,2 @@
 class Person < ActiveRecord::Base
-
 end