+++ spec/support/active_record/person_fuzzy.rb
@@ -1,5 +1,5 @@
 require 'support/active_record/ar_stand_in'
 
-class PersonFuzzy < ARStandIn;
+class PersonFuzzy < ARStandIn
   self.table_name = :people
 end